tested zoom and added properties list to MainActivity screen, sets it to be invisible by default (not seen in other Activities)
