+++ app/src/main/java/org/gwlt/trailapp/SeeMoreActivity.java
@@ -2,6 +2,8 @@ package org.gwlt.trailapp;
 
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
+import android.widget.PopupMenu;
 import android.widget.TextView;
 
 /**
